Refactored getNumValue method
