+++ src/main/resources/plugin-webapp/centaur/app/bottomTabs/options/main.js
@@ -1,9 +1,14 @@
-define(['require', 'angular', './util'], function(require, angular) {
+define(['require', 'angular', './util', '../../common/variableUtil'], function(require, angular) {
 
     /**
      * retrieve the util file containing functions
      */
     var util = require('./util');
+
+    /**
+     * commonUtil containing variable data
+     */
+    var commonUtil = require('../../common/variableUtil');
 
     /**
      * Controller object containing all behavior
@@ -31,7 +36,7 @@ define(['require', 'angular', './util'], function(require, angular) {
         };
 
         $scope.setNumValue = function() {
-            $scope.numValue = util.getNumValue($window.localStorage, procDefId + "_var_num");
+            $scope.numValue = commonUtil.getNumValue($window.localStorage, procDefId + "_var_num");
         };
 
         // get all variable ids for this process
