+++ src/main/resources/plugin-webapp/centaur/app/diagram/instanceVar/main.js
@@ -6,12 +6,11 @@ define(['require', 'angular', './util', '../../bottomTabs/options/util', '../../
      * retrieve the util file containing functions
      */
     var util = require('./util');
-    var commonUtil = require('../../common/variableUtil');
 
     /**
-     * variable containing all ids of overlays created here
+     * common util file for showing variables
      */
-    var optionsUtil = require('../../bottomTabs/options/util');
+    var commonUtil = require('../../common/variableUtil');
 
     /**
      * Overlay object that contains the elements put on the diagram
@@ -28,22 +27,18 @@ define(['require', 'angular', './util', '../../bottomTabs/options/util', '../../
             var overlays = viewer.get('overlays');
             var elementRegistry = viewer.get('elementRegistry');
 
-            // get number of instance variables to show
-            commonUtil.numValue = optionsUtil.getNumValue($window.localStorage, procDefId + "_var_num");
-
             // add the activity variable elements to the overlay
             util.addActivityElements($window, $http, elementRegistry, processDiagram, overlays, Uri, commonUtil);
 
             // subscribe to any broadcast variables options change
             $rootScope.$on("cockpit.plugin.centaur:options:variable-change", function() {
-                util.addActivityElements($window, $http, elementRegistry, processDiagram, overlays, Uri, commonUtil) });
+                util.addActivityElements($window, $http, elementRegistry, processDiagram, overlays, Uri, commonUtil)
+            });
 
             // subscribe to any broadcast variable number changes
             $rootScope.$on("cockpit.plugin.centaur:options:var-num-change", function() {
-                // get number of instance variables to show
-                commonUtil.numValue = optionsUtil.getNumValue($window.localStorage, procDefId + "_var_num");
-
-                util.addActivityElements($window, $http, elementRegistry, processDiagram, overlays, Uri, commonUtil) });
+                util.addActivityElements($window, $http, elementRegistry, processDiagram, overlays, Uri, commonUtil)
+            });
         }
     ];
