+++ src/main/resources/plugin-webapp/centaur/app/plugin.css
@@ -50,7 +50,7 @@
 
 .variableTextSmall {
     font-size: 7px;
-    width: 150px;
+    width: 130px;
     display: flex;
     flex-direction: column;
     border: 1px solid transparent;
